+++ lib/json_web_token.ex
@@ -22,7 +22,7 @@ defmodule JsonWebToken do
   def sign(claims, options), do: Jwt.sign(claims, options)
 
   @doc """
-  Return a JWT claims map if the JWT signature does verify, or an "Invalid" string otherwise
+  Return a JWT claims map if the JWT signature does verify, or a tuple, {:error, "invalid"} otherwise
 
   ## Example
       iex> jwt = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJqb2UiLCJodHRwOi8vZXhhbXBsZS5jb20vaXNfcm9vdCI6dHJ1ZSwiZXhwIjoxMzAwODE5MzgwfQ.Ktfu3EdLz0SpuTIMpMoRZMtZsCATWJHeDEBGrsZE6LI"