+++ test/json_web_token/algorithm/rsa_util_test.exs
@@ -4,14 +4,16 @@ defmodule JsonWebToken.Algorithm.RsaUtilTest do
   alias JsonWebToken.Algorithm.Rsa
   alias JsonWebToken.Algorithm.RsaUtil
 
+  @path_to_keys "test/fixtures/rsa"
+
   test "private_key" do
-    key = RsaUtil.private_key
+    key = RsaUtil.private_key(@path_to_keys, "private_key.pem")
     assert length(key) == 3
     assert byte_size(Rsa.modulus key) == 261
   end
 
   test "public_key" do
-    key = RsaUtil.public_key
+    key = RsaUtil.public_key(@path_to_keys, "public_key.pem")
     assert length(key) == 2
     assert byte_size(Rsa.modulus key) == 261
   end